Port hr_evaluation to 9.0
