+++ hr_evaluation/README.rst
@@ -1,0 +1,103 @@
+.. image:: https://img.shields.io/badge/licence-AGPL--3-blue.svg
+   :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+   :alt: License: AGPL-3
+
+==============================================
+Periodical Employees evaluation and appraisals
+==============================================
+
+By using this application you can maintain the motivational process
+by doing periodical evaluations of your employees' performance.
+The regular assessment of human resources can benefit your people
+as well your organization.
+
+An evaluation plan can be assigned to each employee.
+These plans define the frequency and the way you manage
+your periodic personal evaluations.
+You will be able to define steps and attach interview forms to each step.
+
+Manages several types of evaluations: bottom-up, top-down, self-evaluations
+and the final evaluation by the manager.
+
+Key Features
+------------
+* Ability to create employees evaluations.
+* An evaluation can be created by an employee for subordinates,
+  juniors as well as his manager.
+* The evaluation is done according to a plan in which various surveys
+  can be created. Each survey can be answered by a particular level
+  in the employees hierarchy. The final review and evaluation is done
+  by the manager.
+* Every evaluation filled by employees can be viewed in a PDF form.
+* Interview Requests are generated automatically by Odoo
+  according to employees evaluation plans. Each user receives automatic emails
+  and requests to perform a periodical evaluation of their colleagues.
+
+
+.. figure:: path/to/local/image.png
+   :alt: alternative description
+   :width: 600 px
+
+Usage
+=====
+
+To use this module, you need to:
+
+#. Go to ...
+
+.. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
+   :alt: Try me on Runbot
+   :target: https://runbot.odoo-community.org/runbot/{repo_id}/{branch}
+
+.. repo_id is available in https://github.com/OCA/maintainer-tools/blob/master/tools/repos_with_ids.txt
+.. branch is "8.0" for example
+
+Known issues / Roadmap
+======================
+
+* ...
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues
+<https://github.com/OCA/{project_repo}/issues>`_. In case of trouble, please
+check there if your issue has already been reported. If you spotted it first,
+help us smash it by providing detailed and welcomed feedback.
+
+Credits
+=======
+
+Images
+------
+
+* Odoo Community Association: `Icon <https://github.com/OCA/maintainer-tools/blob/master/template/module/static/description/icon.svg>`_.
+
+Contributors
+------------
+
+* Firstname Lastname <email.address@example.org>
+* Second Person <second.person@example.org>
+
+Funders
+-------
+
+The development of this module has been financially supported by:
+
+* Company 1 name
+* Company 2 name
+
+Maintainer
+----------
+
+.. image:: https://odoo-community.org/logo.png
+   :alt: Odoo Community Association
+   :target: https://odoo-community.org
+
+This module is maintained by the OCA.
+
+OCA, or the Odoo Community Association, is a nonprofit organization whose
+mission is to support the collaborative development of Odoo features and
+promote its widespread use.
+
+To contribute to this module, please visit https://odoo-community.org.
