+++ hr_evaluation/__openerp__.py
@@ -1,4 +1,4 @@
-# -*- encoding: utf-8 -*-
+# -*- coding: utf-8 -*-
 ##############################################################################
 #
 #    OpenERP, Open Source Management Solution
@@ -21,31 +21,12 @@
 
 {
     'name': 'Employee Appraisals',
-    'version': '0.1',
-    'author': 'OpenERP SA',
-    'category': 'Human Resources',
-    'sequence': 31,
-    'website': 'https://www.odoo.com/page/appraisal',
     'summary': 'Periodical Evaluations, Appraisals, Surveys',
+    'version': '9.0.1.0.0',
+    'author': 'OpenERP SA, ',
+    'category': 'Human Resources',
+    'website': 'https://github.com/OCA/hr',
     'depends': ['hr', 'calendar', 'survey'],
-    'description': """
-Periodical Employees evaluation and appraisals
-==============================================
-
-By using this application you can maintain the motivational process by doing periodical evaluations of your employees' performance. The regular assessment of human resources can benefit your people as well your organization.
-
-An evaluation plan can be assigned to each employee. These plans define the frequency and the way you manage your periodic personal evaluations. You will be able to define steps and attach interview forms to each step.
-
-Manages several types of evaluations: bottom-up, top-down, self-evaluations and the final evaluation by the manager.
-
-Key Features
-------------
-* Ability to create employees evaluations.
-* An evaluation can be created by an employee for subordinates, juniors as well as his manager.
-* The evaluation is done according to a plan in which various surveys can be created. Each survey can be answered by a particular level in the employees hierarchy. The final review and evaluation is done by the manager.
-* Every evaluation filled by employees can be viewed in a PDF form.
-* Interview Requests are generated automatically by OpenERP according to employees evaluation plans. Each user receives automatic emails and requests to perform a periodical evaluation of their colleagues.
-""",
     "data": [
         'security/ir.model.access.csv',
         'security/hr_evaluation_security.xml',
@@ -56,11 +37,6 @@ Key Features
         'hr_evaluation_installer.xml',
     ],
     "demo": ["hr_evaluation_demo.xml"],
-    # 'test': [
-    #     'test/test_hr_evaluation.yml',
-    #     'test/hr_evalution_demo.yml',
-    # ],
-    'auto_install': False,
-    'installable': False,
+    'installable': True,
     'application': True,
 }
