+++ hr_evaluation/models/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+from . import hr_evaluation
+from . import hr_evaluation_interview
