+++ hr_evaluation/models/hr_evaluation.py
@@ -1,0 +1,301 @@
+# -*- coding: utf-8 -*-
+##############################################################################
+#
+#    OpenERP, Open Source Management Solution
+#    Copyright (C) 2004-2009 Tiny SPRL (<http://tiny.be>).
+#
+#    This program is free software: you can redistribute it and/or modify
+#    it under the terms of the GNU Affero General Public License as
+#    published by the Free Software Foundation, either version 3 of the
+#    License, or (at your option) any later version.
+#
+#    This program is distributed in the hope that it will be useful,
+#    but WITHOUT ANY WARRANTY; without even the implied warranty of
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#    GNU Affero General Public License for more details.
+#
+#    You should have received a copy of the GNU Affero General Public License
+#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#
+##############################################################################
+
+from datetime import datetime
+from dateutil.relativedelta import relativedelta
+from dateutil import parser
+
+from openerp import api, fields, models
+from openerp.tools.translate import _
+
+
+class EvaluationPlan(models.Model):
+    _name = "hr_evaluation.plan"
+    _description = "Appraisal Plan"
+
+    name = fields.Char("Appraisal Plan", required=True)
+    company_id = fields.Many2one(
+        'res.company',
+        'Company',
+        default=lambda s: s.env.user.company_id,
+        required=True)
+    phase_ids = fields.One2many(
+        'hr_evaluation.plan.phase', 'plan_id', 'Appraisal Phases', copy=True)
+    month_first = fields.Integer(
+        'First Appraisal in (months)',
+        default=6,
+        help="This number of months will be used "
+             "to schedule the first evaluation date of the employee "
+             "when selecting an evaluation plan. ")
+    month_next = fields.Integer(
+        'Periodicity of Appraisal (months)',
+        default=12,
+        help="The number of month that depicts "
+             "the delay between each evaluation of this plan "
+             "(after the first one).")
+    active = fields.Boolean('Active', default=True)
+
+
+class EvaluationPlanPhase(models.Model):
+    _name = "hr_evaluation.plan.phase"
+    _description = "Appraisal Plan Phase"
+    _order = "sequence"
+    name = fields.Char("Phase", required=True)
+    sequence = fields.Integer("Sequence", default=1)
+    company_id = fields.Many2one(
+        related='plan_id.company_id',
+        string='Company',
+        store=True,
+        readonly=True)
+    plan_id = fields.Many2one(
+        'hr_evaluation.plan', 'Appraisal Plan', ondelete='cascade')
+    action = fields.Selection(
+        [('top-down', 'Top-Down Appraisal Requests'),
+         ('bottom-up', 'Bottom-Up Appraisal Requests'),
+         ('self', 'Self Appraisal Requests'),
+         ('final', 'Final Interview')],
+        'Action',
+        required=True)
+    survey_id = fields.Many2one(
+        'survey.survey', 'Appraisal Form', required=True)
+    send_answer_manager = fields.Boolean(
+        'All Answers',
+        help="Send all answers to the manager")
+    send_answer_employee = fields.Boolean(
+        'All Answers',
+        help="Send all answers to the employee")
+    send_anonymous_manager = fields.Boolean(
+        'Anonymous Summary',
+        help="Send an anonymous summary to the manager")
+    send_anonymous_employee = fields.Boolean(
+        'Anonymous Summary',
+        help="Send an anonymous summary to the employee")
+    wait = fields.Boolean(
+        'Wait Previous Phases',
+        help="Check this box if you want to wait that all preceding phases "
+             "are finished before launching this phase.")
+    mail_feature = fields.Boolean(
+        'Send mail for this phase',
+        help="Check this box if you want to send mail "
+             "to employees coming under this phase")
+    mail_body = fields.Text(
+        'Email',
+        default=lambda s: _("""
+Date: %(date)s
+
+Dear %(employee_name)s,
+
+I am doing an evaluation regarding %(eval_name)s.
+
+Kindly submit your response.
+
+
+Thanks,
+--
+%(user_signature)s
+
+        """))
+    email_subject = fields.Text(
+        'Subject',
+        default=lambda s: _('''Regarding '''))
+
+
+class Employee(models.Model):
+    _name = "hr.employee"
+    _inherit = "hr.employee"
+
+    def _compute_appraisal_count(self):
+        Eval = self.env['hr.evaluation.interview']
+        for empl in self:
+            empl.appraisal_count = Eval.search_count(
+                [('user_to_review_id', '=', empl.id)])
+
+    evaluation_plan_id = fields.Many2one(
+        'hr_evaluation.plan', 'Appraisal Plan')
+    evaluation_date = fields.Date(
+        'Next Appraisal Date',
+        help="The date of the next appraisal "
+             "is computed by the appraisal plan's dates "
+             "(first appraisal + periodicity).")
+    appraisal_count = fields.Integer(
+        compute=_compute_appraisal_count,
+        string='Appraisal Interviews')
+
+    @api.model
+    def run_employee_evaluation(self):  # cronjob
+        now = parser.parse(datetime.now().strftime('%Y-%m-%d'))
+        obj_evaluation = self.pool.get('hr_evaluation.evaluation')
+        emp_ids = self.search(cr, uid, [('evaluation_plan_id', '<>', False), ('evaluation_date', '=', False)], context=context)
+        for emp in self.browse(cr, uid, emp_ids, context=context):
+            first_date = (now + relativedelta(months=emp.evaluation_plan_id.month_first)).strftime('%Y-%m-%d')
+            self.write(cr, uid, [emp.id], {'evaluation_date': first_date}, context=context)
+
+        emp_ids = self.search(cr, uid, [('evaluation_plan_id', '<>', False), ('evaluation_date', '<=', time.strftime("%Y-%m-%d"))], context=context)
+        for emp in self.browse(cr, uid, emp_ids, context=context):
+            next_date = (now + relativedelta(months=emp.evaluation_plan_id.month_next)).strftime('%Y-%m-%d')
+            self.write(cr, uid, [emp.id], {'evaluation_date': next_date}, context=context)
+            plan_id = obj_evaluation.create(cr, uid, {'employee_id': emp.id, 'plan_id': emp.evaluation_plan_id.id}, context=context)
+            obj_evaluation.button_plan_in_progress(cr, uid, [plan_id], context=context)
+        return True
+
+
+class Evaluation(models.Model):
+    _name = "hr_evaluation.evaluation"
+    _inherit = "mail.thread"
+    _description = "Employee Appraisal"
+    _rec_name = "employee_id"
+    date = fields.Date(
+        "Appraisal Deadline",
+        default=lambda *a: (parser.parse(datetime.now().strftime('%Y-%m-%d')) + relativedelta(months=+1)).strftime('%Y-%m-%d'),
+        required=True,
+        index=True)
+    employee_id  = fields.Many2one('hr.employee', "Employee", required=True)
+    note_summary = fields.Text('Appraisal Summary')
+    note_action = fields.Text(
+        'Action Plan',
+        help="If the evaluation does not meet the expectations, "
+             "you can propose an action plan")
+    rating = fields.selection(
+        [('0', 'Significantly below expectations'),
+         ('1', 'Do not meet expectations'),
+         ('2', 'Meet expectations'),
+         ('3', 'Exceeds expectations'),
+         ('4', 'Significantly exceeds expectations')],
+        "Appreciation",
+        help="This is the appreciation on which the evaluation is summarized.")
+    survey_request_ids = fields.One2many(
+        'hr.evaluation.interview', 'evaluation_id', 'Appraisal Forms')
+    plan_id = fields.Many2one(
+        'hr_evaluation.plan', 'Plan', required=True)
+    state = fields.Selection(
+        [('draft', 'New'),
+         ('cancel', 'Cancelled'),
+         ('wait', 'Plan In Progress'),
+         ('progress', 'Waiting Appreciation'),
+         ('done', 'Done')],
+        'Status',
+        default='draft',
+        required=True,
+        readonly=True,
+        copy=False)
+    date_close = fields.Date('Ending Date', index=True)
+
+    def name_get(self, cr, uid, ids, context=None):
+        if not ids:
+            return []
+        reads = self.browse(cr, uid, ids, context=context)
+        res = []
+        for record in reads:
+            name = record.plan_id.name
+            employee = record.employee_id.name_related
+            res.append((record['id'], name + ' / ' + employee))
+        return res
+
+    def onchange_employee_id(self, cr, uid, ids, employee_id, context=None):
+        vals = {}
+        vals['plan_id'] = False
+        if employee_id:
+            employee_obj = self.pool.get('hr.employee')
+            for employee in employee_obj.browse(cr, uid, [employee_id], context=context):
+                if employee and employee.evaluation_plan_id and employee.evaluation_plan_id.id:
+                    vals.update({'plan_id': employee.evaluation_plan_id.id})
+        return {'value': vals}
+
+    def button_plan_in_progress(self, cr, uid, ids, context=None):
+        hr_eval_inter_obj = self.env('hr.evaluation.interview')
+        for evaluation in self:
+            wait = False
+            for phase in evaluation.plan_id.phase_ids:
+                children = []
+                if phase.action == "bottom-up":
+                    children = evaluation.employee_id.child_ids
+                elif phase.action in ("top-down", "final"):
+                    if evaluation.employee_id.parent_id:
+                        children = [evaluation.employee_id.parent_id]
+                elif phase.action == "self":
+                    children = [evaluation.employee_id]
+
+                for child in children:
+                    int_id = hr_eval_inter_obj.create(cr, uid, {
+                        'evaluation_id': evaluation.id,
+                        'phase_id': phase.id,
+                        'deadline': (parser.parse(datetime.now().strftime('%Y-%m-%d')) + relativedelta(months=+1)).strftime('%Y-%m-%d'),
+                        'user_id': child.user_id.id,
+                    }, context=context)
+                    if phase.wait:
+                        wait = True
+                    if not wait:
+                        hr_eval_inter_obj.survey_req_waiting_answer(cr, uid, [int_id], context=context)
+
+                    if (not wait) and phase.mail_feature:
+                        body = phase.mail_body % {'employee_name': child.name, 'user_signature': child.user_id.signature,
+                            'eval_name': phase.survey_id.title, 'date': time.strftime('%Y-%m-%d'), 'time': time}
+                        sub = phase.email_subject
+                        if child.work_email:
+                            vals = {'state': 'outgoing',
+                                    'subject': sub,
+                                    'body_html': '<pre>%s</pre>' % body,
+                                    'email_to': child.work_email,
+                                    'email_from': evaluation.employee_id.work_email}
+                            self.pool.get('mail.mail').create(cr, uid, vals, context=context)
+
+        self.write(cr, uid, ids, {'state': 'wait'}, context=context)
+        return True
+
+    def button_final_validation(self, cr, uid, ids, context=None):
+        request_obj = self.pool.get('hr.evaluation.interview')
+        self.write(cr, uid, ids, {'state': 'progress'}, context=context)
+        for evaluation in self.browse(cr, uid, ids, context=context):
+            if evaluation.employee_id and evaluation.employee_id.parent_id and evaluation.employee_id.parent_id.user_id:
+                self.message_subscribe_users(cr, uid, [evaluation.id], user_ids=[evaluation.employee_id.parent_id.user_id.id], context=context)
+            if len(evaluation.survey_request_ids) != len(request_obj.search(cr, uid, [('evaluation_id', '=', evaluation.id), ('state', 'in', ['done', 'cancel'])], context=context)):
+                raise osv.except_osv(_('Warning!'), _("You cannot change state, because some appraisal forms have not been completed."))
+        return True
+
+    def button_done(self, cr, uid, ids, context=None):
+        self.write(cr, uid, ids, {'state': 'done', 'date_close': time.strftime('%Y-%m-%d')}, context=context)
+        return True
+
+    def button_cancel(self, cr, uid, ids, context=None):
+        interview_obj = self.pool.get('hr.evaluation.interview')
+        evaluation = self.browse(cr, uid, ids[0], context)
+        interview_obj.survey_req_cancel(cr, uid, [r.id for r in evaluation.survey_request_ids])
+        self.write(cr, uid, ids, {'state': 'cancel'}, context=context)
+        return True
+
+    def button_draft(self, cr, uid, ids, context=None):
+        self.write(cr, uid, ids, {'state': 'draft'}, context=context)
+        return True
+
+    def write(self, cr, uid, ids, vals, context=None):
+        if vals.get('employee_id'):
+            employee_id = self.pool.get('hr.employee').browse(cr, uid, vals.get('employee_id'), context=context)
+            if employee_id.parent_id and employee_id.parent_id.user_id:
+                vals['message_follower_ids'] = [(4, employee_id.parent_id.user_id.partner_id.id)]
+        if 'date' in vals:
+            new_vals = {'deadline': vals.get('date')}
+            obj_hr_eval_iterview = self.pool.get('hr.evaluation.interview')
+            for evaluation in self.browse(cr, uid, ids, context=context):
+                for survey_req in evaluation.survey_request_ids:
+                    obj_hr_eval_iterview.write(cr, uid, [survey_req.id], new_vals, context=context)
+        return super(hr_evaluation, self).write(cr, uid, ids, vals, context=context)
+
+
