--- hr_evaluation/test/hr_evalution_demo.yml
@@ -1,6 +1,0 @@
-# -
-#   !record {model: hr.employee, id: hr.employee_fp, view: False}:
-#     evaluation_plan_id: hr_evaluation_plan_managersevaluationplan0
-# -
-#   !record {model: hr_evaluation.evaluation, id: hr_evaluation_evaluation_0, view: False}:
-#     plan_id: hr_evaluation.hr_evaluation_plan_managersevaluationplan0
